moved GridCollections component to views subfolder
