+++ resources/js/components/views/grid-collections.tsx
